test(auth): split provider launch and fallback tests (#1389)
